+++ test/python/qobj/test_pulse_converter.py
@@ -115,7 +115,7 @@ class TestInstructionToQobjConverter(QiskitTestCase):
     def test_snapshot(self):
         """Test converted qobj from SnapShot."""
         converter = InstructionToQobjConverter(PulseQobjInstruction, meas_level=2)
-        instruction = Snapshot(name='label', snap_type='type')
+        instruction = Snapshot(label='label', snapshot_type='type')
 
         valid_qobj = PulseQobjInstruction(
             name='snapshot',
@@ -193,7 +193,7 @@ class TestQobjToInstructionConverter(QiskitTestCase):
 
     def test_snapshot(self):
         """Test converted qobj from SnapShot."""
-        cmd = Snapshot(name='label', snap_type='type')
+        cmd = Snapshot(label='label', snapshot_type='type')
         instruction = cmd << 10
 
         qobj = PulseQobjInstruction(name='snapshot', t0=10, label='label', type='type')
@@ -354,7 +354,7 @@ class TestInstructionToQobjConverterWithDeviceSpecification(QiskitTestCase):
     def test_snapshot(self):
         """Test converted qobj from SnapShot."""
         converter = InstructionToQobjConverter(PulseQobjInstruction, meas_level=2)
-        instruction = Snapshot(name='label', snap_type='type')
+        instruction = Snapshot(label='label', snapshot_type='type', name='snapshot')
 
         valid_qobj = PulseQobjInstruction(
             name='snapshot',
@@ -444,7 +444,7 @@ class TestQobjToInstructionConverterWithDeviceSpecification(QiskitTestCase):
 
     def test_snapshot(self):
         """Test converted qobj from SnapShot."""
-        cmd = Snapshot(name='label', snap_type='type')
+        cmd = Snapshot(label='label', snapshot_type='type', name='snapshot')
         instruction = cmd << 10
 
         qobj = PulseQobjInstruction(name='snapshot', t0=10, label='label', type='type')